feat: break out accumulate
